requirements.txt
python-dotenv~=0.15.0
httpx~=0.13.3
pyracing~=0.1.0
setuptools~=51.0.0
logdna~=1.5.3
prettytable~=2.0.0
imgkit~=1.0.2
bokeh~=2.2.3
selenium~=3.141.0
pyracing~=0.1.0
tortoise-orm~=0.16.20
aerich~=0.4.4
